backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
python-multipart
pydantic